+++ public/data/main/Gallary.json
@@ -1,6 +1,7 @@
 {
   "newProduct": [
     {
+      "id": 0,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -12,8 +13,8 @@
           "price": 150000
         },
         {
-          "x": 40,
-          "y": 40,
+          "x": 50,
+          "y": 50,
           "en": "DESK",
           "kr": "책상",
           "explain": "원목 책상",
@@ -22,6 +23,7 @@
       ]
     },
     {
+      "id": 1,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -30,13 +32,14 @@
           "en": "CHAIR",
           "kr": "의자",
           "explain": "합판 의자",
-          "price": 40000
+          "price": 50000
         }
       ]
     }
   ],
   "homeOffice": [
     {
+      "id": 2,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -48,8 +51,8 @@
           "price": 150000
         },
         {
-          "x": 40,
-          "y": 40,
+          "x": 50,
+          "y": 50,
           "en": "DESK",
           "kr": "책상",
           "explain": "원목 책상",
@@ -58,6 +61,7 @@
       ]
     },
     {
+      "id": 3,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -66,13 +70,14 @@
           "en": "CHAIR",
           "kr": "의자",
           "explain": "합판 의자",
-          "price": 40000
+          "price": 50000
         }
       ]
     }
   ],
   "homePunishing": [
     {
+      "id": 4,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -84,8 +89,8 @@
           "price": 150000
         },
         {
-          "x": 40,
-          "y": 40,
+          "x": 50,
+          "y": 50,
           "en": "DESK",
           "kr": "책상",
           "explain": "원목 책상",
@@ -94,6 +99,7 @@
       ]
     },
     {
+      "id": 5,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -102,11 +108,12 @@
           "en": "CHAIR",
           "kr": "의자",
           "explain": "합판 의자",
-          "price": 40000
+          "price": 50000
         }
       ]
     },
     {
+      "id": 6,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -118,8 +125,8 @@
           "price": 150000
         },
         {
-          "x": 40,
-          "y": 40,
+          "x": 50,
+          "y": 50,
           "en": "DESK",
           "kr": "책상",
           "explain": "원목 책상",
@@ -128,6 +135,7 @@
       ]
     },
     {
+      "id": 7,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -136,11 +144,12 @@
           "en": "CHAIR",
           "kr": "의자",
           "explain": "합판 의자",
-          "price": 40000
+          "price": 50000
         }
       ]
     },
     {
+      "id": 8,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -152,8 +161,8 @@
           "price": 150000
         },
         {
-          "x": 40,
-          "y": 40,
+          "x": 50,
+          "y": 50,
           "en": "DESK",
           "kr": "책상",
           "explain": "원목 책상",
@@ -162,6 +171,7 @@
       ]
     },
     {
+      "id": 9,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -170,11 +180,12 @@
           "en": "CHAIR",
           "kr": "의자",
           "explain": "합판 의자",
-          "price": 40000
+          "price": 50000
         }
       ]
     },
     {
+      "id": 10,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -186,8 +197,8 @@
           "price": 150000
         },
         {
-          "x": 40,
-          "y": 40,
+          "x": 50,
+          "y": 50,
           "en": "DESK",
           "kr": "책상",
           "explain": "원목 책상",
@@ -196,6 +207,7 @@
       ]
     },
     {
+      "id": 11,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -204,11 +216,12 @@
           "en": "CHAIR",
           "kr": "의자",
           "explain": "합판 의자",
-          "price": 40000
+          "price": 50000
         }
       ]
     },
     {
+      "id": 12,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -220,8 +233,8 @@
           "price": 150000
         },
         {
-          "x": 40,
-          "y": 40,
+          "x": 50,
+          "y": 50,
           "en": "DESK",
           "kr": "책상",
           "explain": "원목 책상",
@@ -230,6 +243,7 @@
       ]
     },
     {
+      "id": 13,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -238,11 +252,12 @@
           "en": "CHAIR",
           "kr": "의자",
           "explain": "합판 의자",
-          "price": 40000
+          "price": 50000
         }
       ]
     },
     {
+      "id": 14,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -254,8 +269,8 @@
           "price": 150000
         },
         {
-          "x": 40,
-          "y": 40,
+          "x": 50,
+          "y": 50,
           "en": "DESK",
           "kr": "책상",
           "explain": "원목 책상",
@@ -264,6 +279,7 @@
       ]
     },
     {
+      "id": 15,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -272,11 +288,12 @@
           "en": "CHAIR",
           "kr": "의자",
           "explain": "합판 의자",
-          "price": 40000
+          "price": 50000
         }
       ]
     },
     {
+      "id": 16,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -288,8 +305,8 @@
           "price": 150000
         },
         {
-          "x": 40,
-          "y": 40,
+          "x": 50,
+          "y": 50,
           "en": "DESK",
           "kr": "책상",
           "explain": "원목 책상",
@@ -298,6 +315,7 @@
       ]
     },
     {
+      "id": 17,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -306,11 +324,12 @@
           "en": "CHAIR",
           "kr": "의자",
           "explain": "합판 의자",
-          "price": 40000
+          "price": 50000
         }
       ]
     },
     {
+      "id": 18,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -322,8 +341,8 @@
           "price": 150000
         },
         {
-          "x": 40,
-          "y": 40,
+          "x": 50,
+          "y": 50,
           "en": "DESK",
           "kr": "책상",
           "explain": "원목 책상",
@@ -332,6 +351,7 @@
       ]
     },
     {
+      "id": 19,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -340,11 +360,12 @@
           "en": "CHAIR",
           "kr": "의자",
           "explain": "합판 의자",
-          "price": 40000
+          "price": 50000
         }
       ]
     },
     {
+      "id": 20,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -356,8 +377,8 @@
           "price": 150000
         },
         {
-          "x": 40,
-          "y": 40,
+          "x": 50,
+          "y": 50,
           "en": "DESK",
           "kr": "책상",
           "explain": "원목 책상",
@@ -366,6 +387,7 @@
       ]
     },
     {
+      "id": 21,
       "img": "../../image/main/drawer.jpeg",
       "spot": [
         {
@@ -374,7 +396,7 @@
           "en": "CHAIR",
           "kr": "의자",
           "explain": "합판 의자",
-          "price": 40000
+          "price": 50000
         }
       ]
     }
