+++ public/data/main/Recommend.json
@@ -1,5 +1,6 @@
 [
   {
+    "id": 0,
     "img": "../../image/main/banner.jpg",
     "forign": "TAGGVALLMO",
     "korean": "타그발모",
@@ -12,6 +13,7 @@
     "grade": 4
   },
   {
+    "id": 1,
     "img": "../../image/main/banner.jpg",
     "forign": "Billy",
     "korean": "빌리",
@@ -24,6 +26,7 @@
     "grade": 3
   },
   {
+    "id": 2,
     "img": "../../image/main/banner.jpg",
     "forign": "TAGGVALLMO",
     "korean": "타그발모",
@@ -36,6 +39,7 @@
     "grade": 4
   },
   {
+    "id": 3,
     "img": "../../image/main/banner.jpg",
     "forign": "TAGGVALLMO",
     "korean": "타그발모",
@@ -48,6 +52,72 @@
     "grade": 4
   },
   {
+    "id": 4,
+    "img": "../../image/main/banner.jpg",
+    "forign": "TAGGVALLMO",
+    "korean": "타그발모",
+    "info": "매트리스 커버",
+    "width": 90,
+    "height": 200,
+    "zIndex": 0,
+    "unit": "cm",
+    "price": 5000,
+    "grade": 4
+  },
+  {
+    "id": 5,
+    "img": "../../image/main/banner.jpg",
+    "forign": "TAGGVALLMO",
+    "korean": "타그발모",
+    "info": "매트리스 커버",
+    "width": 90,
+    "height": 200,
+    "zIndex": 0,
+    "unit": "cm",
+    "price": 5000,
+    "grade": 4
+  },
+  {
+    "id": 6,
+    "img": "../../image/main/banner.jpg",
+    "forign": "TAGGVALLMO",
+    "korean": "타그발모",
+    "info": "매트리스 커버",
+    "width": 90,
+    "height": 200,
+    "zIndex": 0,
+    "unit": "cm",
+    "price": 5000,
+    "grade": 4
+  },
+  {
+    "id": 7,
+    "img": "../../image/main/banner.jpg",
+    "forign": "TAGGVALLMO",
+    "korean": "타그발모",
+    "info": "매트리스 커버",
+    "width": 90,
+    "height": 200,
+    "zIndex": 0,
+    "unit": "cm",
+    "price": 5000,
+    "grade": 4
+  },
+  {
+    "id": 8,
+    "img": "../../image/main/banner.jpg",
+    "forign": "TAGGVALLMO",
+    "korean": "타그발모",
+    "info": "매트리스 커버",
+    "width": 90,
+    "height": 200,
+    "zIndex": 0,
+    "unit": "cm",
+    "price": 5000,
+    "grade": 4
+  },
+  {
+    "id": 9,
     "img": "../../image/main/banner.jpg",
     "forign": "TAGGVALLMO",
     "korean": "타그발모",
