+++ src/Pages/Main/Gallary/Gallary.scss
@@ -29,28 +29,23 @@
   }
 }
 
-// spot top, bottom 값 퍼센티지로 인라인 부여하기
 .Spot {
   @include flex(center, center);
   flex-direction: column;
   position: absolute;
-  z-index: 100;
   text-decoration: none;
 
   .box {
     @include flex();
     gap: 20px;
-    margin-bottom: 8px;
+    margin-top: 8px;
     padding: 15px;
     z-index: 1000;
+    opacity: 0;
+    transition: 0.4s;
     color: black;
     background-color: white;
     border: 1px solid rgb(201, 201, 201);
-  }
-
-  .text * {
-    display: block;
-    font-weight: bold;
   }
 
   .text {
@@ -58,24 +53,50 @@
     flex-direction: column;
     gap: 5px;
 
+    .isNew {
+      font-weight: bold;
+      color: #ca5008;
+    }
+
+    .name {
+      font-weight: bold;
+    }
+
     .explain {
       font-size: 14px;
-      font-weight: normal;
       color: $gray;
     }
 
     .price {
       margin-top: 10px;
+      font-weight: bold;
     }
   }
 
-  .spotBotton {
-    padding: 15px;
-    border: 2px solid gray;
+  button {
+    box-sizing: border-box;
+    padding: 16px;
+    z-index: 500;
+    border: 0px solid transparent;
     border-radius: 50px;
     background-color: rgb(65, 65, 65);
-    opacity: 0.5;
+    opacity: 0.1;
+    transition: 0.4s;
     cursor: pointer;
+  }
+
+  // button:hover
+
+  button:hover + .box {
+    opacity: 1;
+    transition: 0.6s;
+  }
+
+  button:hover {
+    padding: 15px;
+    border: 2px solid rgb(197, 197, 197);
+    opacity: 0.3;
+    transition: 0.6s;
   }
 }
