+++ src/Pages/Main/Gallary/Spot.js
@@ -3,20 +3,26 @@ import { Link } from 'react-router-dom';
 
 class Spot extends React.Component {
   render() {
-    const { x, y, en, kr, explain, price } = this.props.item;
-    const spot = { bottom: `${x}%`, left: `${y}%` };
+    const { id, item } = this.props;
+    const { x, y, en, kr, explain, price } = item;
+
     return (
-      <Link className="Spot" style={spot} to="">
+      <Link
+        key={id}
+        className="Spot"
+        style={{ bottom: `${x}%`, left: `${y}%` }}
+        to=""
+      >
+        <button></button>
         <div className="box">
           <div className="text">
-            <span style={{ color: 'orange' }}>New</span>
+            <span className="isNew">New</span>
             <span className="name">{`${en} ${kr}`}</span>
             <span className="explain">{explain}</span>
             <span className="price">{`₩ ${price}`}</span>
           </div>
           <i className="fas fa-chevron-right"></i>
         </div>
-        <button className="spotBotton"></button>
       </Link>
     );
   }
