+++ src/Pages/Main/Gallary/Teaser.js
@@ -4,13 +4,16 @@ import Spot from './Spot';
 
 class Teaser extends React.Component {
   render() {
-    const { img, spot } = this.props.items;
+    const { items } = this.props;
+    const { img, spot } = items;
     return (
       <div className="Teaser">
         <Link to="">
           <img src={img} alt={`${img}_img`} />
-          {spot && spot.map(item => <Spot item={item} />)}
         </Link>
+        {spot.map((item, idx) => (
+          <Spot key={idx} item={item} />
+        ))}
       </div>
     );
   }
