+++ src/Pages/Main/Recommend/Recommend.scss
@@ -1,10 +1,9 @@
-@import '../../Styles/variable.scss';
+@import '../../../Styles/variable.scss';
 
 .Recommend {
   margin-bottom: 64px;
 
   * {
-    text-decoration: none;
     color: black;
   }
 
@@ -16,17 +15,16 @@
 
   ul {
     @include flex();
+    flex-wrap: nowrap;
     overflow: scroll;
 
-    img {
-      width: 100%;
+    li {
+      margin-right: 20px;
     }
-  }
 
-  li {
-    @include flex();
-    flex-direction: column;
-    margin-right: 20px;
+    img {
+      max-width: 25rem;
+    }
   }
 
   .text {
