+++ src/Styles/common.scss
@@ -1,0 +1,3 @@
+* {
+  text-decoration: none;
+}
